+++ Userland/Libraries/LibURL/CMakeLists.txt
@@ -4,3 +4,4 @@ set(SOURCES
 )
 
 serenity_lib(LibURL url)
+target_link_libraries(LibURL PRIVATE LibUnicode)
