+++ Userland/Libraries/LibUnicode/CMakeLists.txt
@@ -10,12 +10,10 @@ set(SOURCES
     Segmentation.cpp
     String.cpp
     UnicodeUtils.cpp
-    URL.cpp
     ${UNICODE_DATA_SOURCES}
 )
 set(GENERATED_SOURCES ${CURRENT_LIB_GENERATED})
 
 serenity_lib(LibUnicode unicode)
-target_link_libraries(LibUnicode PRIVATE LibURL)
 
 target_compile_definitions(LibUnicode PRIVATE ENABLE_UNICODE_DATA=$<BOOL:${ENABLE_UNICODE_DATABASE_DOWNLOAD}>)
